chore: unclutter package files
